requirements.txt
black==23.11.0
blessed==1.20.0
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
chalice==1.29.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
inquirer==2.10.1
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
pyrate-limiter==2.10.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
requests-ratelimiter==0.4.2
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.10
